Separate index.js into components and add some styles
