+++ package.json
@@ -30,8 +30,6 @@
     "html-webpack-plugin": "^2.29.0",
     "image-webpack-loader": "^3.3.1",
     "node-sass": "^4.5.3",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
     "rimraf": "^2.6.1",
     "sass-loader": "^6.0.6",
     "style-loader": "^0.18.2",
@@ -39,6 +37,9 @@
     "webpack-dev-server": "^2.5.1"
   },
   "dependencies": {
-    "prop-types": "^15.5.10"
+    "font-awesome": "^4.7.0",
+    "prop-types": "^15.5.10",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1"
   }
 }
