+++ src/static_src/font-awesome/fonts/fontawesome-webfont.woff2
