+++ src/static_src/font-awesome/scss/_fixed-width.scss
@@ -1,0 +1,6 @@
+// Fixed Width Icons
+// -------------------------
+.#{$fa-css-prefix}-fw {
+  width: (18em / 14);
+  text-align: center;
+}
