+++ src/static_src/font-awesome/scss/_screen-reader.scss
@@ -1,0 +1,5 @@
+// Screen Readers
+// -------------------------
+
+.sr-only { @include sr-only(); }
+.sr-only-focusable { @include sr-only-focusable(); }
