+++ src/static_src/sass/style.scss
@@ -1,5 +1,7 @@
 @import "reset";
 @import "base";
+$fa-font-path: "~font-awesome/fonts";
+@import '~font-awesome/scss/font-awesome.scss';
 
 $main-color: #139575;
 $hover-color: #4CC519;
@@ -9,6 +11,7 @@ $hover-color: #4CC519;
   color: #fff;
   background: $main-color;
   padding: 4rem 0;
+  font-weight: 300;
 }
 
 .avatar {
@@ -126,30 +129,11 @@ $hover-color: #4CC519;
 .professional-info__block {
   line-height: 1.5;
   font-size: 1.1rem;
-
-  .professional-info__button {
-      display: inline-block;
-
-      &:hover,
-      &:visited,
-      &:active,
-      &:link {
-          text-decoration: none;
-          cursor: pointer;
-      }
-
-      span {
-          background-color: $main-color;
-      }
-  }
 }
 
 .preview {
     display: inline-block;
-    margin-bottom: 2rem;
-    margin-right: 1rem;
-    border: 1px solid $main-color;
-    width: 70%;
+    width: 100%;
     height: 15rem;
     line-height: 15rem;
     position: relative;
@@ -197,7 +181,35 @@ $hover-color: #4CC519;
 
 .project__block {
   width: 90%;
-  padding-bottom: 2rem;
+  padding: 5px;
+  margin-bottom: 2rem;
+  border: 1px solid $main-color;
+  border-radius: 4px;
+}
+
+.project__name {
+    width: 100%;
+    height: 30px;
+    line-height: 30px;
+    text-align: center;
+    background-color: $main-color;
+    color: #fff;
+}
+
+.project__image {
+    width: 100%;
+}
+
+.project__description {
+    line-height: 1.2;
+    font-size: 15px;
+    padding-top: 1rem;
+    padding-bottom: 1rem;
+}
+
+.project__link {
+    font-size: 15px;
+    padding-bottom: 1rem;
 }
 
 .link {
@@ -214,6 +226,29 @@ $hover-color: #4CC519;
   right: 5%;
   font-size: 2rem;
   color: #fff;
+}
+
+.label {
+    display: inline-block;
+    max-width: 90px;
+    height: 25px;
+    line-height: 25px;
+    padding-right: 5px;
+    padding-left: 5px;
+    margin-right: 8px;
+    border-radius: 4px;
+    background-color: #17bb93;
+    color: #fff;
+    font-size: 14px;
+    font-weight: 300;
+}
+
+.projects__grid {
+    display: grid;
+    width: 100%;
+	grid-template-columns: 1fr;
+  	align-items: start;
+	justify-content: space-between;
 }
 
 @media screen and (min-width: 700px) {
@@ -233,8 +268,11 @@ $hover-color: #4CC519;
   }
 
   .preview {
-    width: 18rem;
     height: 18rem;
     line-height: 18rem;
+  }
+
+  .projects__grid {
+      grid-template-columns: 1fr 1fr;
   }
 }
