+++ webpack.config.js
@@ -35,7 +35,18 @@ module.exports = {
                     'file-loader?name=images/[name].[ext]',
                     'image-webpack-loader'
                 ]
-            }
+            },
+            {
+                 test: /.(ttf|otf|eot|svg|woff(2)?)(\?[a-z0-9]+)?$/,
+                 use: [{
+                     loader: 'file-loader',
+                     options: {
+                         name: '[name].[ext]',
+                         outputPath: 'fonts/',    // where the fonts will go
+                         publicPath: '../'       // override the default path
+                     }
+                 }]
+             }
         ]
     },
     devServer: {
